Modularize for better readability
