+++ test/fixtures/stats.actual
@@ -1,0 +1,1 @@
+{ "type": "Feature", "properties": { "carmen:addressnumber": [ null, [ 100, "200a" ], null ], "carmen:intersections": [ null, null, [ "Main St" ] ], "carmen:rangetype": "tiger" }, "geometry": { "type": "GeometryCollection", "geometries": [ { "type": "MultiLineString", "coordinates": [ [ [ -76.99769139289856, 38.886757140695906 ], [ -76.99419379234314, 38.888218585687234 ] ] ] }, { "type": "MultiPoint", "coordinates": [ [ -76.99553489685059, 38.887024378597644 ], [ -76.99553489685059, 38.888026511775855 ] ] }, { "type": "MultiPoint", "coordinates": [ [ -76.99499845504761, 38.88787619270037 ] ] } ] } }
